menu daily

+++ app/src/main/java/com/team/eatcleanapp/domain/model/NutritionInfo.kt
@@ -1,21 +1,21 @@
 package com.team.eatcleanapp.domain.model
 
 data class NutritionInfo(
-    val totalCalories : Int,
-    val consumedCalories : Int,
+    val totalCalories : Double,
+    val consumedCalories : Double,
 ) {
     // calo con lai
-    val remainingCalories : Int
+    val remainingCalories : Double
         get() = totalCalories - consumedCalories
 
     val isOverLimit: Boolean
         get() = consumedCalories > totalCalories
 
-    val overCalories: Int
-        get() = if (isOverLimit) consumedCalories - totalCalories else 0
+    val overCalories: Double
+        get() = if (isOverLimit) consumedCalories - totalCalories else 0.0
 
-    val remainingNonNegative: Int
-        get() = remainingCalories.coerceAtLeast(0)
+    val remainingNonNegative: Double
+        get() = remainingCalories.coerceAtLeast(0.0)
 
     val progress: Float
         get() = if (totalCalories <= 0) 0f
